backend/requirements.txt
# Core FastAPI Framework
fastapi==0.109.2
uvicorn==0.27.1

# Configuration & Environment
python-dotenv==1.0.0
PyYAML==6.0.1

# HTTP Requests
requests==2.31.0
aiohttp==3.9.3

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# # Development & Testing
# pytest==7.4.3
# pytest-cov==4.1.0
# httpx==0.26.0

# # Code Quality & Formatting
# black==23.3.0
# flake8==6.1.0
# pycodestyle==2.11.1
# autopep8==2.0.4
# mypy==1.7.1